+++ JKEngine/Archer_Arrow_Bye.cpp
@@ -31,7 +31,7 @@ namespace jk
 	{		
 		if (_effect_On == true)
 		{
-			_Lifetime += Time::DeltaTime();
+			_Lifetime += static_cast<float>(Time::DeltaTime());
 			if (_Lifetime >= 7)
 			{
 				at->PlayAnimation(L"ByeEND", true);
