+++ JKEngine/Monster_Hp_Bar.cpp
@@ -22,7 +22,7 @@ namespace jk
 			if (_Type == 1)
 			{
 				if (_target_point < _CurrentHp)				
-					_CurrentHp -= 250*Time::DeltaTime();
+					_CurrentHp -= 250.f * static_cast<float>(Time::DeltaTime());
 				else
 				{
 					_CurrentHp = _target_point;
