+++ JKEngine/Monster_LionWizard.cpp
@@ -93,8 +93,8 @@ namespace jk
 			scene->AddGameObject(eLayerType::Monster, Monster_DamegeHp);
 			Monster_DamegeHp->SetName(L"warrior_hp_bar");
 			Transform* hp_tr = Monster_DamegeHp->GetComponent<Transform>();
-			hp_tr->SetPosition(Vector3(pos.x, pos.y + 50, pos.z - 1.5));
-			hp_tr->SetScale(44, 3, 0);
+			hp_tr->SetPosition(Vector3(pos.x, pos.y + 50.f, pos.z - 1.5f));
+			hp_tr->SetScale(44.f, 3.f, 0.f);
 			Monster_DamegeHp->Set_Max_Hp(_MaxHp);
 			Monster_DamegeHp->Set_Current_Hp(_MaxHp);
 			Monster_DamegeHp->Set_Type(1);
@@ -612,7 +612,7 @@ namespace jk
 		std::uniform_int_distribution<int> distribution(0, 1);
 
 
-		_time += Time::DeltaTime();
+		_time += static_cast<float>(Time::DeltaTime());
 
 		if (_time > 2.5f)
 		{
@@ -704,7 +704,7 @@ namespace jk
 	}
 	void Monster_LionWizard::hit()
 	{
-		_attacktime += Time::DeltaTime();
+		_attacktime += static_cast<float>(Time::DeltaTime());
 		if (_attacktime >= 1)
 		{
 			_state = LionWizard_State::Idle;
@@ -718,7 +718,7 @@ namespace jk
 	}
 	void Monster_LionWizard::dead()
 	{
-		_attacktime += Time::DeltaTime();
+		_attacktime += static_cast<float>(Time::DeltaTime());
 		if (_attacktime >= 1)
 		{
 			_attacktime = 0;
@@ -731,7 +731,7 @@ namespace jk
 	{
 		if (_teleportCheck == false)
 		{
-			_attacktime += Time::DeltaTime();
+			_attacktime += static_cast<float>(Time::DeltaTime());
 			if (_attacktime > 0.15f)
 			{
 				if (mDir == 1)
@@ -778,7 +778,7 @@ namespace jk
 	{
 		if (_hit_particle == true)
 		{
-			_particletime += Time::DeltaTime();
+			_particletime += static_cast<float>(Time::DeltaTime());
 			if (_particletime > 0.5)
 			{
 				Hit_Particle->SetState(eState::Paused);
@@ -790,19 +790,19 @@ namespace jk
 	void Monster_LionWizard::Hpcontrol()
 	{
 		Transform* hp_tr = Monster_Hp->GetComponent<Transform>();
-		hp_tr->SetPosition(Vector3(pos.x, pos.y - 55, pos.z - 2));
+		hp_tr->SetPosition(Vector3(pos.x, pos.y - 55.f, pos.z - 2.f));
 
 		Transform* hpdamege_tr = Monster_DamegeHp->GetComponent<Transform>();
-		hpdamege_tr->SetPosition(Vector3(pos.x, pos.y - 55, pos.z - 1.5));
+		hpdamege_tr->SetPosition(Vector3(pos.x, pos.y - 55.f, pos.z - 1.5f));
 
 		Transform* hpfr_tr = Hpbar_Frame->GetComponent<Transform>();
-		hpfr_tr->SetPosition(Vector3(pos.x, pos.y - 55, pos.z - 1));
+		hpfr_tr->SetPosition(Vector3(pos.x, pos.y - 55.f, pos.z - 1.f));
 
 		if (_Hp_control == true)
 		{
 			if (Monster_DamegeHp->Get_Switch() == true)
 			{
-				_Hp_time += Time::DeltaTime();
+				_Hp_time += static_cast<float>(Time::DeltaTime());
 				if (_Hp_time > 2)
 				{
 					Hpbar_Frame->SetState(eState::Paused);
